[Chore #157033716] Make app consit of units
