+++ app/book.py
@@ -1,0 +1,99 @@
+from flask_restful import Resource
+import random
+
+from .models import Book
+from .user import token_required
+from .parsers import add_book_parser, get_parser, edit_book_parser
+
+
+class AddBook(Resource):
+    """
+    Contains all the methods to add book, list all books
+    """
+    @token_required
+    def post(self, current_user):
+        """Post method to allow addition of book"""
+        args = add_book_parser.parse_args()
+        book_title = args['book_title']
+        authors = args['authors']
+        year = args['year']
+        copies = args['copies']
+        if not book_title or not authors:
+            return {"Message": "Please fill all the details."}, 400
+        book_copies = 0
+        while book_copies < copies:
+            book_copies += 1
+            new_book = Book(book_id=random.randint(1111, 9999), book_title=book_title, authors=authors,
+                            year=year, copies=book_copies)
+            new_book.save_book()
+            result = new_book.book_serializer()
+        return {"Message": "The book was added successfully.", "Book Added": result}, 201
+
+    def get(self):
+        """Get method to get all books"""
+        args = get_parser.parse_args()
+        page = args['page']
+        limit = args['limit']
+        books = Book.query.paginate(page=page, per_page=limit)
+        all_books = books.items
+        num_results = books.total
+        total_pages = books.pages
+        current_page = books.page
+        has_next_page = books.has_next
+        has_prev_page = books.has_prev
+        prev_num = books.prev_num
+        next_num = books.next_num
+        if not all_books:
+            return {"Message": "Books not found"}, 404
+        results = [book.book_serializer() for book in all_books]
+        return {
+            "Total results": num_results,
+            "Total Pages": total_pages,
+            "Current page": current_page,
+            "All books": results,
+            "Previous page": prev_num,
+            "Next page": next_num
+               }, 200
+
+
+class SingleBook(Resource):
+
+    """
+    Contains all activities of a single book, including editing, getting and removing a book.
+    """
+    @token_required
+    def put(self, current_user, book_id):
+        """Put method to edit already existing book"""
+        args = edit_book_parser.parse_args()
+        get_book = Book.query.filter_by(book_id=book_id).first()
+        book_title = args['book_title']
+        authors = args['authors']
+        year = args['year']
+        copies = args['copies']
+        if not get_book:
+            return {"The book is not found"}, 404
+        if get_book:
+            get_book.book_title = book_title
+            get_book.authors = authors
+            get_book.year = year
+            get_book.copies = copies
+            get_book.update_book()
+            edited_book = get_book.book_serializer()
+            return {"Success": edited_book}, 200
+
+    @token_required
+    def delete(self, current_user, book_id):
+        """Delete method to delete a single book"""
+        get_book_id = Book.query.filter_by(book_id=book_id).first()
+        if get_book_id:
+            get_book_id.delete_book()
+            return {"Message": "The book was deleted successfully."}
+        return {"Error": "Book not found."}, 404
+
+    def get(self, book_id):
+        """Get method for a single book"""
+        book = Book.query.filter_by(book_id=book_id).first()
+        if book:
+            result = book.book_serializer()
+            return {"Book": result}, 200
+        return {"Error": "Book not found."}, 404
